remove deprecated SlackClientServer
