+++ engine/apps/slack/models/slack_usergroup.py
@@ -9,8 +9,7 @@ from django.db.models import JSONField
 from django.utils import timezone
 
 from apps.api.permissions import RBACPermission
-from apps.slack.slack_client import SlackClientWithErrorHandling
-from apps.slack.slack_client.exceptions import SlackAPIException
+from apps.slack.client import SlackAPIException, SlackClientWithErrorHandling
 from apps.user_management.models.user import User
 from common.public_primary_keys import generate_public_primary_key, increase_public_primary_key_length
