+++ engine/apps/slack/scenarios/scenario_step.py
@@ -3,7 +3,7 @@ import logging
 import typing
 
 from apps.slack.alert_group_slack_service import AlertGroupSlackService
-from apps.slack.slack_client import SlackClientWithErrorHandling
+from apps.slack.client import SlackClientWithErrorHandling
 
 if typing.TYPE_CHECKING:
     from apps.slack.models import SlackTeamIdentity, SlackUserIdentity
