--- engine/apps/slack/slack_client/__init__.py
@@ -1,1 +1,0 @@
-from .slack_client import SlackClientWithErrorHandling  # noqa: F401
